requirements.txt
pydantic>=1.8.0
requests>=2.25.0
python-dateutil>=2.8.0
typing-extensions>=4.0.0